changed func input to decimal not a double

+++ CSharp-Technology-ADVANCED/Labs/05FunctionalProgramming-Lab/04AddVAT/Program.cs
@@ -7,9 +7,9 @@ namespace _04AddVAT
     {
         static void Main(string[] args)
         {
-            Func<double, string> VAT = x => (x * 1.2).ToString("F2");
+            Func<decimal, string> VAT = x => (x * 1.2M).ToString("F2");
 
-            Console.WriteLine(String.Join(Environment.NewLine, Console.ReadLine().Split(", ", StringSplitOptions.RemoveEmptyEntries).Select(double.Parse).Select(VAT)));
+            Console.WriteLine(String.Join(Environment.NewLine, Console.ReadLine().Split(", ", StringSplitOptions.RemoveEmptyEntries).Select(decimal.Parse).Select(VAT)));
                 
 
             //var prices = Console.ReadLine().Split(new string[] { ", " }, StringSplitOptions.RemoveEmptyEntries).Select(double.Parse).Select(x => x * 1.2).ToArray();
